+++ apis_core/apis_entities/templatetags/apis_entities.py
@@ -4,7 +4,6 @@ from django.contrib.contenttypes.models import ContentType
 from apis_core.apis_entities.forms import EntitiesMergeForm
 from apis_core.apis_entities.models import AbstractEntity
 from apis_core.apis_entities.utils import get_entity_classes
-from apis_core.utils.helpers import triple_sidebar
 
 register = template.Library()
 
@@ -49,12 +48,6 @@ def entities_verbose_name_plural_listview_url():
         for entity in get_entity_classes()
     }
     return sorted(ret.items())
-
-
-@register.simple_tag(takes_context=True)
-def object_relations(context, detail=True):
-    obj = context["object"]
-    return triple_sidebar(obj, context["request"], detail)
 
 
 @register.simple_tag(takes_context=True)
