+++ apis_core/apis_relations/templates/apis_core/apis_entities/abstractentity_detail.html
@@ -1,6 +1,6 @@
 {% extends "apis_core/apis_entities/abstractentity_detail.html" %}
 {% load django_tables2 %}
-{% load apis_entities %}
+{% load apis_relations %}
 
 {% block col-one %}
   <div class="card">
