+++ apis_core/apis_relations/templates/apis_core/apis_entities/abstractentity_form.html
@@ -1,6 +1,6 @@
 {% extends "apis_core/apis_entities/abstractentity_form.html" %}
 {% load core %}
-{% load apis_entities %}
+{% load apis_relations %}
 
 {% block col-one %}
   {{ block.super }}
