+++ apis_core/apis_relations/templates/apis_relations/entity_relations_multiple_cards.html
@@ -1,5 +1,5 @@
 {% load django_tables2 %}
-{% load apis_entities %}
+{% load apis_relations %}
 {% object_relations detail=False as object_relations %}
 {% for obj in object_relations %}
   <div class="card card-default mb-2">
